refactor(common-json): JsonSchema as SPI interface, impl + JsonNode in internal
Addresses PR review:
- JsonSchemaDiagnostic is now a record.
- JsonSchema becomes a public interface in the exported package; the
  validator implementation moves to internal.JsonSchemaImpl and JsonNode
  moves to the internal package alongside it, keeping the implementation
  out of the SPI surface.
